refactor: Restructure UI components
