+++ Front-end/src/app/components/ui/cart/cart-erase-product-dialog.tsx
