+++ Front-end/src/app/components/ui/cart/clear-cart-dialog.tsx
