+++ Front-end/src/app/components/ui/product/add-new-brand-dialog.tsx
