+++ Front-end/src/app/components/ui/product/mobile-product.images.tsx
