+++ Front-end/src/app/components/ui/product/product-layout-toggle.tsx
